requirements.txt
crewai>=0.80.0
crewai-tools>=0.20.0
uvicorn>=0.30.0
fastapi>=0.111.0
websockets>=12.0
pyyaml>=6.0
python-dotenv>=1.0
slack-sdk>=3.27.0
httpx>=0.27.0
anthropic>=0.28.0
